move rand func into test_utils
